+++ apps/docs/app/docs/[...slug]/page.tsx
@@ -1,10 +1,15 @@
 import { Edit } from 'lucide-react';
 import type { Metadata } from 'next';
-import { Card, Cards } from 'fumadocs-ui/components/card';
-import { DocsPage, DocsBody } from 'fumadocs-ui/page';
+import {
+  DocsPage,
+  DocsBody,
+  DocsTitle,
+  DocsDescription,
+  DocsCategory,
+} from 'fumadocs-ui/page';
 import { notFound } from 'next/navigation';
-import { type ReactNode } from 'react';
-import { utils, type Page } from '@/utils/source';
+import { Fragment } from 'react';
+import { utils } from '@/utils/source';
 import { createMetadata } from '@/utils/metadata';
 import Preview from '@/components/preview';
 import { cn } from '@/utils/cn';
@@ -51,56 +56,27 @@ export default function Page({
       full={page.data.full}
       tableOfContent={{
         footer,
+        style: 'clerk',
+        single: true,
       }}
       tableOfContentPopover={{ footer }}
     >
-      <h1 className="text-3xl font-bold text-fd-foreground sm:text-4xl">
-        {page.data.title}
-      </h1>
-      {page.data.description ? (
-        <p className="mb-8 text-lg text-fd-muted-foreground">
-          {page.data.description}
-        </p>
-      ) : null}
+      <DocsTitle>{page.data.title}</DocsTitle>
+      <DocsDescription>{page.data.description}</DocsDescription>
       <DocsBody>
         {preview && preview in Preview ? Preview[preview] : null}
         <page.data.exports.default
           components={{
             HeadlessOnly:
-              params.slug[0] === 'headless'
-                ? ({ children }: { children: ReactNode }) => children
-                : () => undefined,
-            UIOnly:
-              params.slug[0] === 'ui'
-                ? ({ children }: { children: ReactNode }) => children
-                : () => undefined,
+              params.slug[0] === 'headless' ? Fragment : () => undefined,
+            UIOnly: params.slug[0] === 'ui' ? Fragment : () => undefined,
           }}
         />
-        {page.data.index ? <Category page={page} /> : null}
+        {page.data.index ? (
+          <DocsCategory page={page} pages={utils.getPages()} />
+        ) : null}
       </DocsBody>
     </DocsPage>
-  );
-}
-
-function Category({ page }: { page: Page }): React.ReactElement {
-  const filtered = utils
-    .getPages()
-    .filter(
-      (item) =>
-        item.file.dirname === page.file.dirname && item.file.name !== 'index',
-    );
-
-  return (
-    <Cards>
-      {filtered.map((item) => (
-        <Card
-          key={item.url}
-          title={item.data.title}
-          description={item.data.description ?? 'No Description'}
-          href={item.url}
-        />
-      ))}
-    </Cards>
   );
 }